+++ tests/testthat/test-get_predicted.R
@@ -741,3 +741,30 @@ test_that("get_predicted works with brms-Wiener", {
   )
   expect_true(model_info(m)$is_wiener)
 })
+
+
+test_that("get_predicted works with chisq.test", {
+  cat_xtbl <- structure(
+    c(38L, 59L, 54L, 49L),
+    dim = c(2L, 2L),
+    dimnames = list(
+      c("no", "yes"),
+      c("hiphop", "salsa")
+    ),
+    class = "table"
+  )
+
+  cat_chi <- chisq.test(cat_xtbl)
+  out <- get_predicted(cat_chi)
+  expect_identical(
+    out,
+    structure(
+      c(44.62, 52.38, 47.38, 55.62),
+      dim = c(2L, 2L),
+      dimnames = list(
+        c("no", "yes"),
+        c("hiphop", "salsa")
+      )
+    )
+  )
+})